Removed useless components from ducks
