+++ src/ducks/GainersAndLosers/GainersAndLosersTable.js
@@ -1,69 +1,63 @@
 import React from 'react'
 import ReactTable from 'react-table'
-import 'react-table/react-table.css'
+import moment from 'moment'
+import { graphql } from 'react-apollo'
 import getColumns from './gainers-and-losers-table-columns'
-import { fetchGainersOrLosers } from './actions'
-import { connect } from 'react-redux'
-import { bindActionCreators } from 'redux'
+import { TOP_SOCIAL_GAINERS_LOSERS_GQL } from './gainersLosersGQL'
+import { getTimeRangeByDuration } from '../../utils/utils'
+import 'react-table/react-table.css'
 
-class GainersLosersTable extends React.Component {
-  componentDidMount () {
-    const { status, fetchGainersOrLosers, timeWindow } = this.props
-    fetchGainersOrLosers({ status, timeWindow })
-  }
+const GainersLosersTable = ({
+  isLoading,
+  timeWindow,
+  topSocialGainersLosers
+}) => (
+  <ReactTable
+    loading={isLoading}
+    showPaginationTop={false}
+    showPaginationBottom
+    defaultPageSize={10}
+    getTheadThProps={() => ({ style: { textAlign: 'center' } })}
+    getTrProps={() => ({ style: { textAlign: 'center' } })}
+    getTdProps={() => ({
+      style: {
+        display: 'flex',
+        justifyContent: 'center',
+        alignItems: 'center'
+      }
+    })}
+    getTbodyProps={() => ({ style: { overflow: 'hidden' } })}
+    pageSizeOptions={[5, 10, 20, 25, 50, 100]}
+    sortable={false}
+    loadingText='Loading...'
+    resizable
+    data={topSocialGainersLosers}
+    columns={getColumns({ timeWindow })}
+  />
+)
 
-  componentWillReceiveProps ({ status, timeWindow }) {
-    if (status !== this.props.status || timeWindow !== this.props.timeWindow) {
-      this.props.fetchGainersOrLosers({
-        status,
-        timeWindow
-      })
+const withApollo = graphql(TOP_SOCIAL_GAINERS_LOSERS_GQL, {
+  options: ({ status, timeWindow }) => ({
+    variables: {
+      status,
+      size: 10000,
+      ...getTimeRangeByDuration(timeWindow)
+    }
+  }),
+  props: ({ data: { topSocialGainersLosers = [] } }) => {
+    const topGainersDatetime = (topSocialGainersLosers || []).map(
+      ({ datetime }) => moment(datetime)
+    )
+    const recentRecordTime = moment.max(topGainersDatetime)
+    const { projects } =
+      (topSocialGainersLosers || []).find(({ datetime }) =>
+        moment(datetime).isSame(recentRecordTime)
+      ) || {}
+
+    return {
+      topSocialGainersLosers: projects
     }
   }
-
-  render () {
-    const { gainersAndLosers, isLoading, timeWindow } = this.props
-    return (
-      <ReactTable
-        loading={isLoading}
-        showPaginationTop={false}
-        showPaginationBottom
-        defaultPageSize={10}
-        getTheadThProps={() => ({ style: { textAlign: 'center' } })}
-        getTrProps={() => ({ style: { textAlign: 'center' } })}
-        getTdProps={() => ({
-          style: {
-            display: 'flex',
-            justifyContent: 'center',
-            alignItems: 'center'
-          }
-        })}
-        getTbodyProps={() => ({ style: { overflow: 'hidden' } })}
-        pageSizeOptions={[5, 10, 20, 25, 50, 100]}
-        sortable={false}
-        loadingText='Loading...'
-        resizable
-        data={gainersAndLosers}
-        columns={getColumns({ timeWindow })}
-      />
-    )
-  }
-}
-
-const mapStateToProps = ({ gainersLosers }) => ({
-  isLoading: gainersLosers.isLoading,
-  gainersAndLosers: gainersLosers.gainersAndLosers
 })
 
-const mapDispatchToProps = dispatch =>
-  bindActionCreators(
-    {
-      fetchGainersOrLosers
-    },
-    dispatch
-  )
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(GainersLosersTable)
+export default withApollo(GainersLosersTable)
