+++ src/epics/rootEpics.js
@@ -58,7 +58,6 @@ import {
   insightDraftPublishEpic
 } from '../pages/Insights/insightDraftEpic'
 import { likesEpic } from '../components/Like/likesEpic'
-import { fetchGainersEpic } from '../ducks/GainersAndLosers/gainersLosersEpic'
 
 export default combineEpics(
   handleOffline,
@@ -115,7 +114,5 @@ export default combineEpics(
   insightDraftUpdateEpic,
   insightDraftPublishEpic,
   // likes
-  likesEpic,
-  // gainers and losers
-  fetchGainersEpic
+  likesEpic
 )
