+++ src/reducers/rootReducers.js
@@ -27,9 +27,6 @@ import hypedTrends, {
 import wordCloud, {
   initialState as initialWordCloud
 } from './../components/WordCloud/reducers'
-import gainersLosers, {
-  initialState as initialGainersLosers
-} from '../ducks/GainersAndLosers/reducers'
 import socialVolume, {
   initialState as initialSocialVolume
 } from './../components/SocialVolumeWidget/reducers'
@@ -55,8 +52,7 @@ export const intitialState = {
   wordCloud: initialWordCloud,
   socialVolume: initialSocialVolume,
   accountUi: initialAccountUiState,
-  insightDraft: initialInsightDraft,
-  gainersLosers: initialGainersLosers
+  insightDraft: initialInsightDraft
 }
 
 export default combineReducers({
@@ -73,6 +69,5 @@ export default combineReducers({
   wordCloud,
   socialVolume,
   accountUi,
-  insightDraft,
-  gainersLosers
+  insightDraft
 })
